Reorganize unit test assemblies
