+++ src/SkynetServer/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 ﻿using System;
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("SkynetServer.Protocol.Tests")] 
+[assembly: InternalsVisibleTo("SkynetServer.Tests")]
