+++ tests/SkynetServer.Shared.Tests/DatabaseHelperTests.cs
@@ -1,6 +1,5 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SkynetServer.Configuration;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SkynetServer.Database;
 using SkynetServer.Database.Entities;
 using SkynetServer.Model;
 using SkynetServer.Threading;
@@ -10,26 +9,11 @@ using System.Security.Cryptography;
 using System.Threading.Tasks;
 using Wiry.Base32;
 
-namespace SkynetServer.Database.Tests
+namespace SkynetServer.Shared.Tests
 {
     [TestClass]
     public class DatabaseHelperTests
     {
-        [AssemblyInitialize]
-        public static void AssemblyInitialize(TestContext context)
-        {
-            var configuration = new ConfigurationBuilder()
-                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-                .Build();
-
-            DatabaseContext.ConnectionString = configuration.Get<SkynetOptions>().DatabaseOptions.ConnectionString;
-
-            using (DatabaseContext ctx = new DatabaseContext())
-            {
-                ctx.Database.EnsureCreated();
-            }
-        }
-
         [TestMethod]
         public async Task TestAddAccount()
         {
