+++ tests/SkynetServer.Tests/PacketTests.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace SkynetServer.Protocol.Tests
+namespace SkynetServer.Tests
 {
     [TestClass]
     public class PacketTests
