+++ tests/SkynetServer.Tests/SkynetServer.Tests.csproj
